Switch from FHIR DSTU3 to FHIR R4.
Underlying libraries updated plus all Java code references. Some work on
the Java code was necessary to fix various models to comply with R4.

com.github.ben-manes.caffeine was added to build.gradle because there is
a problem with the HAPI FHIR library where it requires it use but it is
not pulled down as a dependency.