MedicationStatement Output US Core STU3 Fixes
Mainly work is to add medicationCodeableConcept & provide the dates
(dateAsserted and effectivePeriod/start,end). It turns out that the
medicationCodableConcept support was already there, but the VistA logic
for getting RxNorm codes was not correct. This is not fixed; and a fall
back has been provided to code using VUIDs or Drug IENs in this order if
an RxNorm connection cannot be made.

Java Changes:
HcConstants.java: Add urn:dxc:vista:drug as a possible coding system.
MedicationParser.java:
- result.setId had a off by one error. Now fixed.
- Vista now sends 3 dates: asserted; start; end. Parser updated to use these.
- Parse coding system as RxNorm, VUID or Drug in that order.

M Changes:
SYNDHP48:
- Return asserted date; start date; end date to Java. Dates are mapped
to fields as follows:
Outpatient:
52,1     = Issue Date      = Asserted Date
52,22    = Fill Date       = Start Date
52,26    = Expiration Date = End Date
Inpatient:
55.06,27 = Order Date      = Asserted Date
55.06,10 = Start Date
55.06,34 = End Date
- Fall back to VUID, Drug IEN if RxNorm could not be obtained.
SYNDHPUTL:
- Fix RxNorm get logic. Try VA asserteed map in 50.68 first and then the
full RxNorm data in the ETS package.